Configured google strategy passport

+++ controllers/api/v1/posts_api.js
@@ -23,8 +23,8 @@ module.exports.destroy = async function (req, res) {
 
     try {
         
-        let post = await Post.findById(req.params.id);
-    //   if (post.user == req.user.id) {
+      let post = await Post.findById(req.params.id);
+      if (post.user == req.user.id) {
         post.remove();
   
         await Comment.deleteMany({ post: req.params.id });
@@ -33,10 +33,11 @@ module.exports.destroy = async function (req, res) {
             message: "Post and associated comments deleted successfully!"
         });
   
-    //   } else {
-    //     req.flash('error','You cannot delete this post');
-    //     return res.redirect('back');
-    //   }
+      } else {
+        return res.json(401, {
+          message: "You cannot delete this post"
+        });
+      }
   
     } catch (err) {
       return res.json(500, {
